src/models/phi3_vision/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.41.0
accelerate>=0.30.0
Pillow>=10.0.0
opencv-python>=4.9.0